- rpm.org: More assorted int -> size_t uses.

+++ rpmio/macro.c
@@ -1789,7 +1789,7 @@ static int XpoptParseArgvString(const char * s, int * argcPtr, const char *** ar
     int argvAlloced = POPT_ARGV_ARRAY_GROW_DELTA;
     const char ** argv = malloc(sizeof(*argv) * argvAlloced);
     int argc = 0;
-    int buflen = strlen(s) + 1;
+    size_t buflen = strlen(s) + 1;
     char * buf = memset(alloca(buflen), 0, buflen);
     int rc = POPT_ERROR_MALLOC;
